reports/containervulns/requirements.txt
python-dotenv~=0.15
requests~=2.25
configparser~=5.0
bleach~=4.0
urllib3~=1.26
xlsxwriter~=3.0.9
pandas~=1.3.5
laceworksdk~=1.9.0
